refactor(debug_server): extract /mode endpoint family (refs #332)
Sixth per-family extract from debug_server.c (after K144 #336, OTA
#340, codec #341, voice #342 + the #343 hotfix that healed main).
Now branched from a clean main, no stacking — avoids the conflict-
marker issue that bit #342.

Endpoint owned by this PR:
  POST /mode?m=0|1|2|3|4&model=...  — switch active voice mode

Plus the small async LVGL-thread refresh helper (`async_refresh_mode_badge`)
that's only used by this handler — moves with it.

## Diff

  main/CMakeLists.txt       |   2 +-
  main/debug_server.c       |  80 +-----  (handler + helper + URI def + register call)
  main/debug_server_mode.h  |  21 +++ (new)
  main/debug_server_mode.c  | 100 +++ (new)

debug_server.c after this:  ~3859 LOC (was 3938 before this PR).
Cumulative shrinkage from origin/main (pre-Wave-23b 4520 LOC):
  4520 → 3859 = **−661 LOC, 14.6 %**.

## Hardware verification

build:  clean (only debug_server_mode.c compiled)
flash:  OK to /dev/ttyACM0, boot uptime ~30 s

POST /mode?m=2&model=anthropic/claude-3.5-haiku → {
   voice_mode:2, mode_name:"cloud",
   llm_model:"anthropic/claude-3.5-haiku",
   voice_connected:true, status:"applied"
}
POST /mode?m=0  → {mode:"local", status:"applied"}

story_smoke (14 steps): 14/14 PASS in 180 s.
   Step 3 "Force voice mode = Local (0)" exercises /mode via the new
   module; full Local-mode chat round-trip in step 7 confirms the
   mode actually applied (Dragon-side STT/LLM/TTS all swap accordingly).
