+++ main/CMakeLists.txt
@@ -34,7 +34,7 @@ else()
              "sdcard.c" "wifi.c"
              "voice.c" "voice_codec.c" "voice_video.c" "mode_manager.c"
              "debug_server.c" "debug_server_m5.c" "debug_server_ota.c"
-             "debug_server_codec.c" "debug_server_voice.c"
+             "debug_server_codec.c" "debug_server_voice.c" "debug_server_mode.c"
              "debug_obs.c" "pool_probe.c" "heap_watchdog.c"
              "service_registry.c" "service_storage.c" "service_display.c"
              "service_audio.c" "service_network.c" "service_dragon.c"
